+++ packages/api-client/src/libs/send-request/decode-buffer.ts
@@ -1,5 +1,6 @@
-import { isTextMediaType } from '@/views/Request/consts'
 import MimeTypeParser from 'whatwg-mimetype'
+
+import { isTextMediaType } from '@/views/Request/consts'
 
 // TODO: This should return `unknown` to acknowledge we don't know type, shouldn't it?
 /** Decode the buffer according to its content-type */