+++ packages/api-reference/src/components/Content/Schema/Schema.vue
@@ -21,6 +21,8 @@ const {
   compact,
   noncollapsible = false,
   hideHeading,
+  hideReadOnly,
+  hideWriteOnly,
   additionalProperties,
   hideModelNames = false,
   discriminator,
@@ -37,6 +39,10 @@ const {
   noncollapsible?: boolean
   /** Hide the heading */
   hideHeading?: boolean
+  /** Hide read-only properties */
+  hideReadOnly?: boolean
+  /** Hide write-only properties */
+  hideWriteOnly?: boolean
   /** Show a special one way toggle for additional properties, also has a top border when open */
   additionalProperties?: boolean
   /** Hide model names in type display */
@@ -182,6 +188,8 @@ const handleClick = (e: MouseEvent) => noncollapsible && e.stopPropagation()
             :discriminator
             :hideHeading="hideHeading"
             :hideModelNames="hideModelNames"
+            :hideReadOnly="hideReadOnly"
+            :hideWriteOnly="hideWriteOnly"
             :level="level + 1"
             :schema="schema" />
 
@@ -193,6 +201,8 @@ const handleClick = (e: MouseEvent) => noncollapsible && e.stopPropagation()
               :compact
               :hideHeading
               :hideModelNames
+              :hideReadOnly="hideReadOnly"
+              :hideWriteOnly="hideWriteOnly"
               :level
               :value="schema" />
           </template>