+++ packages/pre-post-request-scripts/src/index.ts
@@ -1,1 +1,2 @@
+export { ViewLayoutCollapse } from './components/ViewLayout'
 export { postResponseScriptsPlugin } from './plugins/post-response-scripts'