+++ integrations/docker/CHANGELOG.md
@@ -1,5 +1,26 @@
 # @scalarapi/docker-api-reference
 
+## 0.2.32
+
+### Patch Changes
+
+- Updated dependencies [b6a59bd]
+- Updated dependencies [8f2a2f2]
+  - @scalar/api-reference@1.35.6
+
+## 0.2.31
+
+### Patch Changes
+
+- Updated dependencies [4e7e8a9]
+  - @scalar/api-reference@1.35.5
+
+## 0.2.30
+
+### Patch Changes
+
+- @scalar/api-reference@1.35.4
+
 ## 0.2.29
 
 ### Patch Changes