+++ integrations/nuxt/CHANGELOG.md
@@ -1,5 +1,35 @@
 # @scalar/nuxt
 
+## 0.5.13
+
+### Patch Changes
+
+- Updated dependencies [b6a59bd]
+- Updated dependencies [8f2a2f2]
+  - @scalar/api-reference@1.35.6
+  - @scalar/api-client@2.5.33
+  - @scalar/use-hooks@0.2.5
+  - @scalar/workspace-store@0.15.6
+
+## 0.5.12
+
+### Patch Changes
+
+- Updated dependencies [4e7e8a9]
+  - @scalar/api-reference@1.35.5
+  - @scalar/workspace-store@0.15.5
+  - @scalar/api-client@2.5.32
+
+## 0.5.11
+
+### Patch Changes
+
+- Updated dependencies [019a22a]
+- Updated dependencies [f8fc81c]
+  - @scalar/workspace-store@0.15.4
+  - @scalar/api-client@2.5.31
+  - @scalar/api-reference@1.35.4
+
 ## 0.5.10
 
 ### Patch Changes