+++ packages/components/CHANGELOG.md
@@ -1,5 +1,31 @@
 # @scalar/components
 
+## 0.14.34
+
+### Patch Changes
+
+- 498f771: fix(components): default scalar menu links to be <a> tags
+- 8f2a2f2: Bump some dependencies to latest
+- Updated dependencies [bff46e5]
+- Updated dependencies [8f2a2f2]
+  - @scalar/helpers@0.0.11
+  - @scalar/use-hooks@0.2.5
+  - @scalar/themes@0.13.17
+  - @scalar/oas-utils@0.4.29
+  - @scalar/code-highlight@0.1.9
+
+## 0.14.33
+
+### Patch Changes
+
+- @scalar/oas-utils@0.4.28
+
+## 0.14.32
+
+### Patch Changes
+
+- @scalar/oas-utils@0.4.27
+
 ## 0.14.31
 
 ### Patch Changes