+++ packages/oas-utils/CHANGELOG.md
@@ -1,5 +1,33 @@
 # @scalar/oas-utils
 
+## 0.4.29
+
+### Patch Changes
+
+- Updated dependencies [bff46e5]
+- Updated dependencies [8f2a2f2]
+  - @scalar/helpers@0.0.11
+  - @scalar/themes@0.13.17
+  - @scalar/json-magic@0.4.3
+  - @scalar/object-utils@1.2.7
+  - @scalar/workspace-store@0.15.6
+
+## 0.4.28
+
+### Patch Changes
+
+- Updated dependencies [3bd1209]
+- Updated dependencies [1943b99]
+  - @scalar/json-magic@0.4.2
+  - @scalar/workspace-store@0.15.5
+
+## 0.4.27
+
+### Patch Changes
+
+- Updated dependencies [019a22a]
+  - @scalar/workspace-store@0.15.4
+
 ## 0.4.26
 
 ### Patch Changes