+++ packages/openapi-to-markdown/CHANGELOG.md
@@ -1,5 +1,33 @@
 # @scalar/openapi-to-markdown
 
+## 0.2.37
+
+### Patch Changes
+
+- Updated dependencies [498f771]
+- Updated dependencies [bff46e5]
+- Updated dependencies [8f2a2f2]
+  - @scalar/components@0.14.34
+  - @scalar/helpers@0.0.11
+  - @scalar/oas-utils@0.4.29
+  - @scalar/openapi-parser@0.20.6
+
+## 0.2.36
+
+### Patch Changes
+
+- Updated dependencies [39bbc0e]
+  - @scalar/openapi-parser@0.20.5
+  - @scalar/oas-utils@0.4.28
+  - @scalar/components@0.14.33
+
+## 0.2.35
+
+### Patch Changes
+
+- @scalar/oas-utils@0.4.27
+- @scalar/components@0.14.32
+
 ## 0.2.34
 
 ### Patch Changes