+++ packages/use-codemirror/CHANGELOG.md
@@ -1,5 +1,25 @@
 # @scalar/use-codemirror
 
+## 0.12.35
+
+### Patch Changes
+
+- Updated dependencies [498f771]
+- Updated dependencies [8f2a2f2]
+  - @scalar/components@0.14.34
+
+## 0.12.34
+
+### Patch Changes
+
+- @scalar/components@0.14.33
+
+## 0.12.33
+
+### Patch Changes
+
+- @scalar/components@0.14.32
+
 ## 0.12.32
 
 ### Patch Changes